+++ packages/react-components/src/stories/form/StyledModalButton.tsx
@@ -3,18 +3,18 @@ import StyledVerticalStack from '../layout-helpers/StyledVerticalStack';
 import DfxIcon, { IconSize, IconVariant } from '../DfxIcon';
 
 export interface StyledModalButtonProps {
-  label?: string;
   onClick: () => void;
   onBlur: Noop;
+  label?: string;
   value?: string;
   description?: string;
   placeholder: string;
 }
 
 export default function StyledModalButton({
-  label,
   onClick,
   onBlur,
+  label,
   value,
   description,
   placeholder,
@@ -22,7 +22,7 @@ export default function StyledModalButton({
 }: StyledModalButtonProps): JSX.Element {
   return (
     <StyledVerticalStack gap={1}>
-      <label className="text-dfxBlue-800 text-base font-semibold pl-4 text-start">{label}</label>
+      {label && <label className="text-dfxBlue-800 text-base font-semibold pl-4 text-start">{label}</label>}
       <button
         type="button"
         className="flex justify-between border border-dfxGray-400 text-base font-normal rounded-md px-4 py-2 shadow-sm w-full h-[58px]"