+++ packages/react/src/definitions/route.ts
@@ -30,7 +30,6 @@ export enum PaymentLinkPaymentMode {
 
 export enum PaymentStandardType {
   OPEN_CRYPTO_PAY = 'OpenCryptoPay',
-  FRANKENCOIN_PAY = 'FrankencoinPay',
   LIGHTNING_BOLT11 = 'LightningBolt11',
   PAY_TO_ADDRESS = 'PayToAddress',
 }
@@ -59,6 +58,7 @@ export const PaymentLinkBlockchain = {
   BASE: Blockchain.BASE,
   ETHEREUM: Blockchain.ETHEREUM,
   LIGHTNING: Blockchain.LIGHTNING,
+  BITCOIN: Blockchain.BITCOIN,
   MONERO: Blockchain.MONERO,
   OPTIMISM: Blockchain.OPTIMISM,
   POLYGON: Blockchain.POLYGON,