+++ packages/react/src/hooks/kyc.hook.ts
@@ -21,6 +21,8 @@ import {
   TfaLevel,
   KycFile,
   KycStepBase,
+  KycBeneficialData,
+  KycOperationalData,
 } from '../definitions/kyc';
 import { useApi } from './api.hook';
 
@@ -50,6 +52,8 @@ export interface KycInterface {
   setNationalityData: (code: string, url: string, data: KycNationalityData) => Promise<KycStepBase>;
   setFileData: (code: string, url: string, data: KycFileData) => Promise<KycStepBase>;
   setSignatoryPowerData: (code: string, url: string, data: KycSignatoryPowerData) => Promise<KycStepBase>;
+  setBeneficialData: (code: string, url: string, data: KycBeneficialData) => Promise<KycStepBase>;
+  setOperationalData: (code: string, url: string, data: KycOperationalData) => Promise<KycStepBase>;
   getFinancialData: (code: string, url: string, lang?: string) => Promise<KycFinancialQuestions>;
   setFinancialData: (code: string, url: string, data: KycFinancialResponses) => Promise<KycStepBase>;
   getFile: (kycFileId: string) => Promise<KycFile>;
@@ -138,6 +142,14 @@ export function useKyc(): KycInterface {
     return call({ url, code, method: 'PUT', data });
   }
 
+  function setBeneficialData(code: string, url: string, data: KycBeneficialData): Promise<KycStepBase> {
+    return call({ url, code, method: 'PUT', data });
+  }
+
+  function setOperationalData(code: string, url: string, data: KycOperationalData): Promise<KycStepBase> {
+    return call({ url, code, method: 'PUT', data });
+  }
+
   async function getFinancialData(code: string, url: string, lang?: string): Promise<KycFinancialQuestions> {
     lang && (url += `?lang=${lang}`);
     return call({ url, code, method: 'GET' });
@@ -218,6 +230,8 @@ export function useKyc(): KycInterface {
       setNationalityData,
       setFileData,
       setSignatoryPowerData,
+      setBeneficialData,
+      setOperationalData,
       getFinancialData,
       getFile,
       setFinancialData,